chore: fix for assigned validators
The SSV API doesnt return the validator index, so it was displaying the wrong assigned validators.

Quick fix was to display the validator pub keys. Also prepared an endpoint which returns the validator index based on the pub_key, if we want to display that instead of the pub key later.